moved operators to passFilter bas class
