+++ Filter/LowPass.hpp
@@ -1,12 +1,9 @@
 #ifndef LOWPASS_HPP
 #define LOWPASS_HPP
 
-#include "Filter.hpp"
+#include "PassFilter.hpp"
 
-class Bandpass;
-class HighPass;
-
-class LowPass: public Filter 
+class LowPass: public PassFilter 
 {
 public:
   virtual ~LowPass() = default;
@@ -15,16 +12,10 @@ public:
   LowPass(const LowPass&) = default;
   LowPass& operator=(const LowPass&) = default;
   
-  virtual double Frequency() const = 0;
-
-  friend std::unique_ptr<Bandpass> operator+ (
-    std::shared_ptr<LowPass> const lowPass,
-    std::shared_ptr<HighPass> const highPass);
+  virtual double Frequency() const override = 0;
 
 protected:
   LowPass() {};
-  std::unique_ptr<Component> m_firstComponent;
-  std::unique_ptr<Component> m_secondComponent;
     
 };
 #endif //LOWPASS_HPP
